country is added in services

+++ lib/application2/booking_bloc_and_data/bloc/booking_event.dart
@@ -9,8 +9,8 @@ abstract class BookingEvent extends Equatable {
 
 class CreateBookingEvent extends BookingEvent {
   final String serviceId;
-  final int latitude;
-  final int longitude;
+  final double latitude;
+  final double longitude;
   final String scheduledDate;
   final String scheduledTimeStart;
   final String address;
